dont use ballmodel here
