+++ CMakeLists.txt
@@ -300,7 +300,7 @@ set(INTERFACE_SOURCES
         ${PROJECT_SOURCE_DIR}/roboteam_ai/src/interface/TreeVisualizerWidget.cpp
         ${PROJECT_SOURCE_DIR}/roboteam_ai/src/interface/TreeVisualizerWidget.h
         ${PROJECT_SOURCE_DIR}/roboteam_ai/src/interface/RobotsWidget.cpp
-        ${PROJECT_SOURCE_DIR}/roboteam_ai/src/interface/RobotsWidget.h roboteam_ai/src/coach/PassCoach.cpp roboteam_ai/src/coach/PassCoach.h roboteam_ai/src/coach/Ballplacement.cpp roboteam_ai/src/coach/Ballplacement.h)
+        ${PROJECT_SOURCE_DIR}/roboteam_ai/src/interface/RobotsWidget.h roboteam_ai/src/coach/PassCoach.cpp roboteam_ai/src/coach/PassCoach.h roboteam_ai/src/coach/Ballplacement.cpp roboteam_ai/src/coach/Ballplacement.h roboteam_ai/src/coach/GeneralPositionCoach.cpp roboteam_ai/src/coach/GeneralPositionCoach.h)
 
 catkin_package()
 include_directories(${catkin_INCLUDE_DIRS})
