+++ roboteam_ai/src/skills/EnterFormation.cpp
@@ -14,6 +14,8 @@ std::vector<std::shared_ptr<roboteam_msgs::WorldRobot>> EnterFormation::robotsIn
 EnterFormation::EnterFormation(std::string name, bt::Blackboard::Ptr blackboard) : Skill(std::move(name), std::move(blackboard)) {}
 
 void EnterFormation::onInitialize() {
+
+    // add the robot if its not already there.
     for (unsigned long i = 0; i<robotsInFormation.size(); i++) {
         if (robotsInFormation.at(i)->id == robot->id) {
             return;
@@ -21,13 +23,18 @@ void EnterFormation::onInitialize() {
     }
 
     robotsInFormation.push_back(robot);
-  //  robotsInFormationMemory = robotsInFormation.size();
 }
 
 bt::Node::Status EnterFormation::onUpdate() {
+
+    /*
+     * Calculate the target location at least once, and every time when the amount of robots in the formation change.
+     */
     if (robotsInFormationMemory != robotsInFormation.size()) {
         targetLocation = getFormationPosition();
     }
+
+
     auto robotPos = rtt::Vector2(robot->pos);
     Vector2 targetToLookAtLocation = Field::get_their_goal_center();
