+++ roboteam_ai/src/utilities/Coach.h
@@ -25,29 +25,13 @@ class Coach {
     FRIEND_TEST(CoachTest, it_adds_and_removes_defenders);
     FRIEND_TEST(CoachTest, it_adds_and_removes_formationrobots);
 
-
 public:
     static std::map<int, int> defencePairs;
-    static int pickOffensivePassTarget(int selfID, std::string roleName);
-    static int pickDefensivePassTarget(int selfID);
     static int pickOpponentToCover(int selfID);
-    static int pickHarassmentTarget(int selfID);
-
-    static Vector2 getPositionBehindBallToGoal(double distanceBehindBall, bool ourGoal);
-    static Vector2 getPositionBehindBallToRobot(double distanceBehindBall, bool ourRobot, const unsigned int &robotID);
-    static Vector2 getPositionBehindBallToPosition(double distanceBehindBall, const Vector2 &position);
-
-    static std::shared_ptr<roboteam_msgs::WorldRobot> getRobotClosestToPosition(
-            std::vector<roboteam_msgs::WorldRobot> &robots,
-            Vector2 position, bool includeSamePosition);
 
 
-    static bool isRobotBehindBallToGoal(double distanceBehindBall, bool ourGoal, const Vector2 &robotPos);
-    static bool isRobotBehindBallToRobot(double distanceBehindBall, bool ourRobot, const unsigned int &robotID, const Vector2 &robotPosition);
-    static bool isRobotBehindBallToPosition(double distanceBehindBall, const Vector2 &position, const Vector2 &robotPosition);
 
     static std::pair<int, bool> getRobotClosestToBall();
-
     static std::shared_ptr<roboteam_msgs::WorldRobot> getRobotClosestToBall(bool isOurTeam);
     static Vector2 getDemoKeeperGetBallPos(Vector2 ballPos);
 };
